feat: link subsciption with payment and reformat
