--- apps/web/app/auth/signup-creator/final-steps/page.tsx
@@ -1,12 +1,0 @@
-import CreatorFinalSteps from "@/components/Feature/Auth/CreatorFinalSteps";
-import { Main, PageContainer } from "../../signup/styles";
-
-export default function CreatorFinalStepsPage() {
-  return (
-    <PageContainer>
-      <Main>
-        <CreatorFinalSteps />
-      </Main>
-    </PageContainer>
-  );
-}
