+++ apps/web/components/Feature/Subscription/SubscriptionPaymentStep.styles.ts
@@ -1,49 +1,14 @@
 import styled from "styled-components";
-import { media } from "@repo/ui/breakpoints";
 import GenericButton from "@/components/UI/GenericButton";
 import { SIZE, VARIANT } from "@/utils/Constants";
 
-export const Screen = styled.section`
-  min-height: 100vh;
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background: ${({ theme }) => theme.colors.neutral.OFF_WHITE};
-  padding: 24px;
-
-  ${media.mobileLg} {
-    align-items: flex-start;
-    padding: 32px 16px;
-  }
-`;
-
-export const Card = styled.div`
+export const PaymentCard = styled.div`
   width: min(100%, 312px);
-  min-height: 580px;
   display: flex;
   flex-direction: column;
-  align-items: center;
-  background: transparent;
-  padding: 0 0 32px;
-
-  ${media.mobileLg} {
-    padding: 0 0 28px;
-  }
 `;
 
-export const LogoWrap = styled.div`
-  display: flex;
-  justify-content: center;
-  margin-bottom: 30px;
-  transform: translateY(-1px);
-
-  img {
-    display: block;
-  }
-`;
-
-export const Title = styled.h1`
+export const PaymentTitle = styled.h1`
   margin: 0 0 20px;
   color: ${({ theme }) => theme.colors.primary.BLACK};
   text-align: center;
@@ -56,37 +21,8 @@ export const Form = styled.form`
 `;
 
 export const PlanSelectWrap = styled.div`
-  position: relative;
   width: 100%;
   margin-bottom: 42px;
-`;
-
-export const PlanSelect = styled.select`
-  width: 100%;
-  height: 38px;
-  appearance: none;
-  border: 0;
-  border-radius: 10px;
-  background: ${({ theme }) => theme.colors.primary.PALE_GREEN};
-  color: ${({ theme }) => theme.colors.neutral.GRAY};
-  padding: 0 42px 0 16px;
-  outline: none;
-  cursor: pointer;
-  ${({ theme }) => theme.typography.Body_Medium};
-
-  &:focus-visible {
-    outline: 2px solid ${({ theme }) => theme.colors.primary.BLACK};
-    outline-offset: 2px;
-  }
-`;
-
-export const SelectIcon = styled.span`
-  position: absolute;
-  right: 15px;
-  top: 50%;
-  display: inline-flex;
-  transform: translateY(-50%);
-  pointer-events: none;
 `;
 
 export const PaymentMethods = styled.div`
@@ -115,11 +51,6 @@ export const PaymentMethodButton = styled.button<{ $active?: boolean }>`
   gap: 10px;
   cursor: pointer;
   overflow: hidden;
-
-  &:focus-visible {
-    outline: 2px solid ${({ theme }) => theme.colors.primary.BLACK};
-    outline-offset: 2px;
-  }
 `;
 
 export const PaymentCardLogoWrap = styled.span`
