+++ apps/web/locals/da.json
@@ -153,7 +153,6 @@
     }
   },
   "creatorFinalSteps": {
-    "logoAlt": "Kiibee",
     "title": "Sidste trin",
     "plans": {
       "tryKiibee": "Try Kiibee (0 KR)",
