+++ apps/web/locals/en.json
@@ -153,7 +153,6 @@
     }
   },
   "creatorFinalSteps": {
-    "logoAlt": "Kiibee",
     "title": "Final steps",
     "plans": {
       "tryKiibee": "Try Kiibee (0 KR)",
