+++ apps/web/utils/Constants.ts
@@ -75,5 +75,6 @@ export const APPEARANCE_DEFAULT_HEX_COLOR = "#674096";
 export const SUBSCRIPTION_STEP = {
   PLAN: "plan",
   DETAILS: "details",
+  PAYMENT: "payment",
 } as const;
 export const maxReceiptCharacters = 200;
